Google places classes moved to other package.
